pyproject.toml
[project]
name = "dbs-annotator"
description = "An application for annotating clinical programming sessions."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Lucia Poma", email = "lucia.poma@wysscenter.ch" },
    { name = "Matteo Vissani" },
    { name = "Richard Köhler", email = "richard.koehler@wysscenter.ch" },
]
keywords = ["dbs", "annotation", "neuroscience", "deep-brain-stimulation"]
dynamic = ["version"]
classifiers = [
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Environment :: X11 Applications :: Qt",
    "Operating System :: OS Independent",
]
dependencies = [
    "tzdata>=2025.3",
    "pandas>=2.1",
    "python-docx>=1.1",
    "docx2pdf>=0.1.8",
    "pydantic>=2.13.0",
    "pyside6>=6.11.0",
    "matplotlib>=3.5",
    "packaging>=23.0",
]

[project.optional-dependencies]
# Toga/Briefcase build extras. The Qt desktop installers keep using the
# core ``dependencies`` list above; mobile and Toga-desktop builds pull
# these extras so the Qt wheel is never shipped to iOS/Android.
toga = ["toga>=0.5.0"]
mobile = [
    "toga>=0.5.0",
    "reportlab>=4.2",
    "cairosvg>=2.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
    "pytest-timeout>=2.3.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6",
    "pre-commit-uv>=4.2.1",
    "ty>=0.0.29",
    "pandas-stubs>=3.0.0.260204",
    "pyside6-stubs>=6.7.3.0",
    "towncrier>=24.8.0",
    "toga-dummy>=0.5.3",
]
build = ["briefcase>=0.3.26"]
docs = [
    "sphinx>=9.0.4",
    "sphinx-rtd-theme>=3.1.0",
    "sphinx-copybutton>=0.5.2",
    "myst-parser>=4.0.0",
]

[project.scripts]
dbs-annotator = "dbs_annotator.__main__:main"

[project.urls]
Homepage = "https://github.com/Brain-Modulation-Lab/App_ClinicalDBSAnnot"
Repository = "https://github.com/Brain-Modulation-Lab/App_ClinicalDBSAnnot"
Documentation = "https://app-clinicaldbsannot.readthedocs.io/en/latest/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dbs_annotator/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/dbs_annotator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--timeout=3",
    "--cov=dbs_annotator",
    "--cov-report=html",
    "--cov-report=term-missing",
]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "gui: Qt GUI / widget tests",
    "integration: cross-module behavior",
    "slow: longer GUI or lazy-loaded wizard flows; skip locally with pytest -m \"not slow\"",
    "docs_screenshot: optional screenshot generation for docs artifacts",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
omit = ["*/tests/*"]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = [".eggs", ".git", ".hatch", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]

[tool.ruff.lint.per-file-ignores]
"src/dbs_annotator/views/wizard_window.py" = ["E501"]

[tool.codespell]
ignore-words-list = "vas"
skip = "uv.lock,.venv,build,dist,htmlcov,docs/_build,*.svg,*.ico,*.png,*.pdf,*.log,*.min.js"

[tool.doc8]
max-line-length = 100
ignore = ["D004"]
ignore-path = ["docs/_build", "docs/_templates"]

[tool.interrogate]
fail-under = 85
ignore-init-module = true
ignore-magic = true
ignore-private = true
ignore-semiprivate = true
ignore-nested-functions = true
exclude = ["tests", "build", "dist", ".venv", "docs/_build"]
verbose = 0
quiet = false

[tool.uv]
exclude-newer = "1 week"

[tool.towncrier]
package = "dbs_annotator"
package_dir = "src"
directory = "newsfragments"
filename = "CHANGELOG.md"
template = "newsfragments/template.md.jinja"
start_string = "## [Unreleased]"
title_format = "## [{version}] - {project_date}"
issue_format = "#{issue}"
underlines = ["", "", ""]

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "docs"
name = "Documentation"
showcontent = true

[tool.briefcase]
project_name = "DBSAnnotator"
bundle = "ch.wysscenter.dbsannotator"
# Briefcase requires a static version here even when [project] uses dynamic version
version = "0.4.0a1"
license = { file = "LICENSE" }
author = "WyssCenter"
author_email = "lucia.poma@wysscenter.ch"
url = "https://github.com/Brain-Modulation-Lab/App_ClinicalDBSAnnot"

[tool.briefcase.app.dbs_annotator]
formal_name = "DBSAnnotator"
description = "A graphical user interface for annotating DBS clinical programming sessions"
long_description = """
DBS Annotator is a desktop application used to structure and review
deep-brain stimulation programming sessions in a reproducible workflow.

It supports clinicians and researchers by organizing session information,
tracking stimulation settings, and generating standardized exports and reports.
"""
sources = ["src/dbs_annotator", "icons", "styles"]
icon = "icons/logoneutral"
installer_icon = "icons/logoneutral"
requirement_installer_args = [
    "--index-url",
    "https://pypi.org/simple",
    "--constraint",
    "constraints-briefcase.txt",
    "--retries",
    "5",
    "--timeout",
    "60",
    "--no-cache-dir",
]

[tool.briefcase.app.dbs_annotator.macOS]
universal_build = false
# PySide6 6.8+ publishes macOS wheels tagged `macosx_13_0_universal2`. The Briefcase
# support-package Python on macOS advertises a lower platform tag by default, so its
# embedded pip refuses all PySide6 versions newer than 6.7.3 with a misleading
# "No matching distribution" error. Setting `min_os_version = "13.0"` tells Briefcase
# (and the bundled interpreter/installer) to target macOS 13+, aligning platform tags
# with the published wheels. If a compatible support package is not available, Briefcase
# will select the closest one and raise the deployment target accordingly.
min_os_version = "13.0"

# ---------------------------------------------------------------------------
# Mobile Briefcase targets.
#
# These targets intentionally do NOT inherit the Qt ``dependencies`` from the
# top-level [project] block: iOS/Android have no PySide6 wheels. Instead we
# list the Toga backend + the mobile-only runtime stack explicitly. Once
# Phase 4 of the migration completes, the desktop targets move over too and
# the [project] ``dependencies`` list is trimmed to the backend-neutral core.
# ---------------------------------------------------------------------------

[tool.briefcase.app.dbs_annotator.iOS]
requires = [
    "toga-iOS~=0.5.0",
    "std-nslog~=1.0.3",
    "python-docx>=1.1",
    "reportlab>=4.2",
    "pydantic>=2.13.0",
    "packaging>=23.0",
]

[tool.briefcase.app.dbs_annotator.android]
requires = [
    "toga-android~=0.5.0",
    "python-docx>=1.1",
    "reportlab>=4.2",
    "pydantic>=2.13.0",
    "packaging>=23.0",
]

base_theme = "Theme.MaterialComponents.Light.DarkActionBar"

build_gradle_dependencies = [
    "com.google.android.material:material:1.12.0",
    "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0",
    "androidx.constraintlayout:constraintlayout:2.2.0",
    "androidx.coordinatorlayout:coordinatorlayout:1.3.0",
]
